chore: Structural refactor + type hint cleanup.
- Move `SonosState` to its own package
- Move `SonosDriver` to its own package
- Cleanup type hints/doc comments
