+++ drivers/SmartThings/sonos/.gitignore
@@ -1,1 +1,6 @@
 app_key.lua
+.vscode/
+.zed/
+.editorconfig
+.luarc.json
+.stylua.toml
