+++ drivers/SmartThings/sonos/src/api/cmd_handlers.lua
@@ -38,28 +38,32 @@ end
 
 function CapCommandHandlers.handle_play(driver, device, _cmd)
   local payload = {
-    { namespace = "playback", command = "play" }, {}
+    { namespace = "playback", command = "play" },
+    {},
   }
   _do_send_to_group(driver, device, payload)
 end
 
 function CapCommandHandlers.handle_pause(driver, device, _cmd)
   local payload = {
-    { namespace = "playback", command = "pause" }, {}
+    { namespace = "playback", command = "pause" },
+    {},
   }
   _do_send_to_group(driver, device, payload)
 end
 
 function CapCommandHandlers.handle_next_track(driver, device, _cmd)
   local payload = {
-    { namespace = "playback", command = "skipToNextTrack" }, {}
+    { namespace = "playback", command = "skipToNextTrack" },
+    {},
   }
   _do_send_to_group(driver, device, payload)
 end
 
 function CapCommandHandlers.handle_previous_track(driver, device, _cmd)
   local payload = {
-    { namespace = "playback", command = "skipToPreviousTrack" }, {}
+    { namespace = "playback", command = "skipToPreviousTrack" },
+    {},
   }
   _do_send_to_group(driver, device, payload)
 end
@@ -76,7 +80,7 @@ function CapCommandHandlers.handle_set_mute(driver, device, cmd)
   local set_mute = (cmd.args and cmd.args.state == "muted")
   local payload = {
     { namespace = "playerVolume", command = "setMute" },
-    { muted = set_mute }
+    { muted = set_mute },
   }
   _do_send_to_self(driver, device, payload)
 end
@@ -84,7 +88,7 @@ end
 function CapCommandHandlers.handle_volume_up(driver, device, cmd)
   local payload = {
     { namespace = "playerVolume", command = "setRelativeVolume" },
-    { volumeDelta = 5 }
+    { volumeDelta = 5 },
   }
   _do_send_to_self(driver, device, payload)
 end
@@ -92,7 +96,7 @@ end
 function CapCommandHandlers.handle_volume_down(driver, device, cmd)
   local payload = {
     { namespace = "playerVolume", command = "setRelativeVolume" },
-    { volumeDelta = -5 }
+    { volumeDelta = -5 },
   }
   _do_send_to_self(driver, device, payload)
 end
@@ -101,7 +105,7 @@ function CapCommandHandlers.handle_set_volume(driver, device, cmd)
   local new_volume = st_utils.clamp_value(cmd.args.volume, 0, 100)
   local payload = {
     { namespace = "playerVolume", command = "setVolume" },
-    { volume = new_volume }
+    { volume = new_volume },
   }
   _do_send_to_self(driver, device, payload)
 end
@@ -118,7 +122,7 @@ function CapCommandHandlers.handle_group_set_mute(driver, device, cmd)
   local set_mute = (cmd.args and cmd.args.state == "muted")
   local payload = {
     { namespace = "groupVolume", command = "setMute" },
-    { muted = set_mute }
+    { muted = set_mute },
   }
   _do_send_to_group(driver, device, payload)
 end
@@ -126,7 +130,7 @@ end
 function CapCommandHandlers.handle_group_volume_up(driver, device, cmd)
   local payload = {
     { namespace = "groupVolume", command = "setRelativeVolume" },
-    { volumeDelta = 5 }
+    { volumeDelta = 5 },
   }
   _do_send_to_group(driver, device, payload)
 end
@@ -134,7 +138,7 @@ end
 function CapCommandHandlers.handle_group_volume_down(driver, device, cmd)
   local payload = {
     { namespace = "groupVolume", command = "setRelativeVolume" },
-    { volumeDelta = -5 }
+    { volumeDelta = -5 },
   }
   _do_send_to_group(driver, device, payload)
 end
@@ -143,7 +147,7 @@ function CapCommandHandlers.handle_group_set_volume(driver, device, cmd)
   local new_volume = st_utils.clamp_value(cmd.args.groupVolume, 0, 100)
   local payload = {
     { namespace = "groupVolume", command = "setVolume" },
-    { volume = new_volume }
+    { volume = new_volume },
   }
   _do_send_to_group(driver, device, payload)
 end
@@ -154,8 +158,8 @@ function CapCommandHandlers.handle_play_preset(driver, device, cmd)
     {
       favoriteId = cmd.args.presetId,
       playOnCompletion = true,
-      action = (device.preferences[QUEUE_ACTION_PREF] or "APPEND")
-    }
+      action = (device.preferences[QUEUE_ACTION_PREF] or "APPEND"),
+    },
   }
   _do_send_to_group(driver, device, payload)
 end
@@ -167,10 +171,10 @@ function CapCommandHandlers.handle_audio_notification(driver, device, cmd)
       appId = "edge.smartthings.com",
       name = "SmartThings Audio Notification",
       streamUrl = cmd.args.uri,
-    }
+    },
   }
 
-  if type(cmd.args.level) == 'number' then
+  if type(cmd.args.level) == "number" then
     payload[2].volume = cmd.args.level
   end
   _do_send_to_self(driver, device, payload)
