+++ drivers/SmartThings/sonos/src/fields.lua
@@ -1,4 +1,4 @@
---- @module 'fields'
+--- @class Fields
 local Fields = {}
 
 ---@enum SonosPlayerFields
@@ -6,9 +6,14 @@ Fields.SonosPlayerFields = {
   _IS_INIT = "init",
   _IS_SCANNING = "scanning",
   CONNECTION = "conn",
-  HOUSEHOULD_ID = "householdId",
+  UNIQUE_KEY = "unique_key",
+  HOUSEHOLD_ID = "householdId",
   PLAYER_ID = "playerId",
+  GROUP_ID = "groupId",
+  GROUP_ROLE = "groupRole",
+  COORDINATOR_ID = "coordinatorId",
   WSS_URL = "wss_url",
+  REST_URL = "rest_url",
   SW_GEN = "sw_gen",
 }
