+++ drivers/SmartThings/sonos/src/types.lua
@@ -1,32 +1,35 @@
-local handlers = require "api.event_handlers"
-local log = require "log"
-local st_utils = require "st.utils"
+---@meta
 
-local PlayerFields = require "fields".SonosPlayerFields
-
---- @module 'sonos.types'
-local Types = {}
-
---- @enum SonosCapabilities
-Types.SonosCapabilities = {
-  PLAYBACK = "PLAYBACK",                   --- The player can produce audio. You can target it for playback.
-  CLOUD = "CLOUD",                         --- The player can send commands and receive events over the internet.
-  HT_PLAYBACK = "HT_PLAYBACK",             --- The player is a home theater source. It can reproduce the audio from a home theater system, typically delivered by S/PDIF or HDMI.
-  HT_POWER_STATE = "HT_POWER_STATE",       --- The player can control the home theater power state. For example, it can switch a connected TV on or off.
-  AIRPLAY = "AIRPLAY",                     --- The player can host AirPlay streams. This capability is present when the device is advertising AirPlay support.
-  LINE_IN = "LINE_IN",                     --- The player has an analog line-in.
-  AUDIO_CLIP = "AUDIO_CLIP",               ---  The device is capable of playing audio clip notifications.
-  VOICE = "VOICE",                         --- The device supports the voice namespace (not yet implemented by Sonos).
-  SPEAKER_DETECTION = "SPEAKER_DETECTION", --- The component device is capable of detecting connected speaker drivers.
-  FIXED_VOLUME = "FIXED_VOLUME"            --- The device supports fixed volume.
-}
-
+--- @alias UniqueKey string unique identifier composed of the form `${HouseholdId}/${PlayerId}`
 --- @alias PlayerId string
 --- @alias HouseholdId string
 --- @alias GroupId string
 
+--- @alias SonosCapabilities
+---| "PLAYBACK" # The player can produce audio. You can target it for playback.
+---| "CLOUD" # The player can send commands and receive events over the internet.
+---| "HT_PLAYBACK" # The player is a home theater source. It can reproduce the audio from a home theater system, typically delivered by S/PDIF or HDMI.
+---| "HT_POWER_STATE" # The player can control the home theater power state. For example, it can switch a connected TV on or off.
+---| "AIRPLAY" # The player can host AirPlay streams. This capability is present when the device is advertising AirPlay support.
+---| "LINE_IN" # The player has an analog line-in.
+---| "AUDIO_CLIP" #  The device is capable of playing audio clip notifications.
+---| "VOICE" # The device supports the voice namespace (not yet implemented by Sonos).
+---| "SPEAKER_DETECTION" # The component device is capable of detecting connected speaker drivers.
+---| "FIXED_VOLUME" # The device supports fixed volume.
+
+--- @class SonosFeatureInfo
+--- @field public _objectType "feature"
+--- @field public name string
+
+---@class SonosVersionsInfo
+---@field public _objectType "sdkVersions"
+---@field public audioTxProtocol { [1]: integer }
+---@field public trueplaySdk { [1]: string }
+---@field public controlApi { [1]: string }
+
 --- Lua representation of the Sonos `deviceInfo` JSON Object: https://developer.sonos.com/build/control-sonos-players-lan/discover-lan/#deviceInfo-object
 --- @class SonosDeviceInfo
+--- @field public _objectType "deviceInfo"
 --- @field public id PlayerId The playerId. Also known as the deviceId. Used to address Sonos devices in the control API.
 --- @field public primaryDeviceId string Identifies the primary device in bonded sets. Primary devices leave the value blank, which omits the key from the message. The field is expected for secondary devices in stereo pairs and satellites in home theater configurations.
 --- @field public serialNumber string The device serial number printed on the device.
@@ -40,15 +43,20 @@ Types.SonosCapabilities = {
 --- @field public softwareVersion string Stores the software version the player is running.
 --- @field public hwVersion string Stores the hardware version the player is running. The format is: `{vendor}.{model}.{submodel}.{revision}-{region}.`
 --- @field public swGen integer Stores the software generation that the player is running.
+--- @field public versions SonosVersionsInfo
+--- @field public features SonosFeatureInfo[]
 
 --- Lua representation of the Sonos `discoveryInfo` JSON object: https://developer.sonos.com/build/control-sonos-players-lan/discover-lan/#discoveryInfo-object
 --- @class SonosDiscoveryInfo
+--- @field public _objectType "discoveryInfo"
 --- @field public device SonosDeviceInfo The device object. This object presents immutable data that describes a Sonos device. Use this object to uniquely identify any Sonos device. See below for details.
 --- @field public householdId HouseholdId An opaque identifier assigned to the device during registration. This field may be missing prior to registration.
 --- @field public playerId PlayerId The identifier used to address this particular device in the control API.
 --- @field public groupId GroupId The currently assigned groupId, an ephemeral opaque identifier. This value is always correct, including for group members.
 --- @field public websocketUrl string The URL to the WebSocket server. Use this interface to receive real time updates from the device.
 --- @field public restUrl string The base URL for REST commands using the control API. You can use the same format as REST commands sent to the Sonos cloud. Sonos exposes all REST commands relative to this URL.
+--- @field public allowGuestAccess boolean whether guest access is enabled
+--- @field public credentialTypeAllowed string
 
 --- Lua representation of the Sonos LAN API response's header structure
 --- @class SonosResponseHeader
@@ -61,12 +69,60 @@ Types.SonosCapabilities = {
 --- @field public playerId? PlayerId ID of the target if the target was a player. A header will have one of `sessionId`, `groupId`, or `playerId`,
 --- @field public householdId HouseholdId the Household ID
 
---- Lua representation of the Sonos `groups` (note the plural) JSON object: https://developer.sonos.com/reference/control-api/groups/groups/
+--- Lua representation of a Sonos error body
+--- @class SonosErrorResponse
+--- @field public _objectType "globalError"
+--- @field public errorCode "ERROR_NOT_CAPABLE"|"ERROR_COMMAND_FAILED"|"ERROR_INVALID_OBJECT_ID"|"ERROR_INVALID_PARAMETER"|"ERROR_UNSUPPORTED_COMMAND"|"ERROR_UNSUPPORTED_NAMESPACE"|"ERROR_INVALID_SYNTAX"|"ERROR_MISSING_PARAMETERS"|"ERROR_NOT_AUTHORIZED"
+--- @field public wwwAuthenticate string? Used with `errorCode` is `ERROR_NOT_AUTHORIZED`
+--- @field public reason string? Used with `ERROR_MISSING_PARAMETER`
+
+--- Lua representation of the Sonos `groups` (note the plural) JSON object: https://developer.sonos.com/reference/control-api/groups
 --- @class SonosGroupsResponseBody
+--- @field public _objectType "groups"
 --- @field public groups SonosGroupObject[]
 --- @field public players SonosPlayerObject[]
+--- @field public partial boolean will be `true` if the API response excluded players/groups because it considered them "invalid"
 
---- Lua representation of the Sonos `group` (note the singular) JSON object: https://developer.sonos.com/reference/control-api/groups/groups/#group
+--- Lua representation of the Sonos `favorites` (note the plural) JSON object: https://developer.sonos.com/reference/control-api/favorites
+--- @class SonosFavoritesResponseBody
+--- @field public _objectType "favoritesList"
+--- @field public version string
+--- @field public items any[]
+
+--- Lua representation of a Sonos `favorite` (note the singular) JSON object: https://developer.sonos.com/reference/control-api/favorites
+--- @class SonosFavoriteObject
+--- @field public _objectType "favorite"
+--- @field public id string
+--- @field public name string
+--- @field public description string
+--- @field public imageUrl string? deprecated, the `images` array parameter is preferred
+--- @field public images SonosImageObject[]
+--- @field public service SonosServiceObject
+--- @field public resource SonosContentResourceObject
+
+--- @class SonosImageObject
+--- @field public _objectType "image"
+--- @field public url string
+
+--- @class SonosServiceObject
+--- @field public _objectType "service"
+--- @field public name string
+--- @field public id string
+--- @field public images SonosImageObject[]
+
+--- @class SonosContentResourceObject
+--- @field public _objectType "contentResource"
+--- @field public type string
+--- @field public name string
+--- @field public images SonosImageObject[]
+
+--- @class SonosMusicObjectId
+--- @field public _objectType "universalMusicObjectId"
+--- @field public serviceId string
+--- @field public objectId string
+--- @field public accountId string
+
+--- Lua representation of the Sonos `group` (note the singular) JSON object: https://developer.sonos.com/reference/control-api/groups
 --- @class SonosGroupObject
 --- @field public coordinatorId PlayerId
 --- @field public id GroupId
@@ -103,282 +159,19 @@ Types.SonosCapabilities = {
 --- @field public group_id GroupId
 --- @field public group_name string
 --- @field public household_id HouseholdId
+--- @field public player_id PlayerId extracted from the USN
 --- @field public wss_url string
+--- @field public expires_at integer
 
 --- @alias SonosFavorites { id: string, name: string }[]
+--- @alias DiscoCallback fun(dni: string, ssdp_group_info: SonosSSDPInfo, player_info: SonosDiscoveryInfo, group_info: SonosGroupsResponseBody): boolean?
+
 --- @class SonosHousehold
 --- Information on an entire Sonos system ("household"), such as its current groups, list of players, etc.
+---
 --- @field public groups table<GroupId,SonosGroupObject> All of the current groups in the system
 --- @field public players table<PlayerId,SonosPlayerObject> All of the current players in the system
 --- @field public favorites SonosFavorites all of the favorites/presets in the system
-
---- Sonos systems local state
---- @class SonosState
---- @field public get_household fun(self: SonosState, id: HouseholdId): SonosHousehold
---- @field public update_household_info fun(self: SonosState, id: HouseholdId, groups_event: SonosGroupsResponseBody, device: SonosDevice|nil)
---- @field public update_household_favorites fun(self: SonosState, id: HouseholdId, favorites: SonosFavorites)
---- @field public get_group_for_player fun(self: SonosState, household_id: HouseholdId, player_id: PlayerId): GroupId
---- @field public get_coordinator_for_player fun(self: SonosState, household_id: HouseholdId, player_id: PlayerId): PlayerId
---- @field public get_coordinator_for_group fun(self: SonosState, household_id: HouseholdId, group_id: GroupId): PlayerId
---- @field public get_player_for_device fun(self: SonosState, device: SonosDevice): HouseholdId,PlayerId,string
---- @field public get_coordinator_for_device fun(self: SonosState, device: SonosDevice): HouseholdId,PlayerId,string
---- @field public get_group_for_device fun(self: SonosState, device: SonosDevice): HouseholdId,GroupId,string
---- @field public mark_player_as_joined fun(self: SonosState, player_id: PlayerId)
---- @field public mark_player_as_removed fun(self: SonosState, player_id: PlayerId)
---- @field public is_player_joined fun(self: SonosState, household_id_or_dni: HouseholdId|string, player_id?: PlayerId): boolean
-local SonosState = {}
-SonosState.__index = SonosState
-
-function SonosState.new()
-  local ret = setmetatable({}, SonosState)
-
-  local private = {
-    households = {},
-    joined_players = {}
-  }
-
-  ret.mark_player_as_joined = function(self, player_id)
-    log.debug(string.format("Marking Player ID %s as joined", player_id))
-    private.joined_players[player_id] = true
-  end
-
-  ret.mark_player_as_removed = function(self, player_id)
-    log.debug(string.format("Marking Player ID %s as removed", player_id))
-    private.joined_players[player_id] = false
-  end
-
-  ret.is_player_joined = function(self, dni)
-    return private.joined_players[dni] ~= nil and private.joined_players[dni] -- we want a boolean return, not nil
-  end
-
-  ret.get_household = function(self, id)
-    return private.households[id]
-  end
-
-  --- @param self SonosState
-  --- @param id HouseholdId
-  --- @param favorites SonosFavorites
-  ret.update_household_favorites = function(self, id, favorites)
-    local household = private.households[id] or {} --- @type SonosHousehold
-    household.favorites = favorites or {}
-  end
-
-  --- @param self SonosState
-  --- @param id HouseholdId
-  --- @param groups_event SonosGroupsResponseBody
-  --- @param device SonosDevice|nil
-  ret.update_household_info = function(self, id, groups_event, device)
-    log.debug_with({ hub_logs = true },
-      st_utils.stringify_table(
-        {name = (device or { label = "<no device>" }).label, event = groups_event },
-        string.format("Update household info for household %s", id),
-        true
-      )
-    )
-    if device and device.label then
-      log.debug(string.format("Household update triggered by device %s to update capabilities", device.label))
-    end
-    local household = private.households[id] or {}
-    local groups, players = groups_event.groups, groups_event.players
-
-    -- We create these maps to avoid having to constantly iterate
-    -- when performing lookups. We use the various ID's as keys.
-    household.groups = {}
-    household.players = {}
-    household.player_to_group_map = {}
-    household.group_to_coordinator_map = {}
-
-    for _, group in ipairs(groups) do
-      household.groups[group.id] = group
-      household.group_to_coordinator_map[group.id] = group.coordinatorId
-      for _, playerId in ipairs(group.playerIds) do
-        household.player_to_group_map[playerId] = group.id
-      end
-    end
-
-    for _, player in ipairs(players) do
-      household.players[player.id] = player
-    end
-
-    household.player_to_group_map.__newindex = function(_, _, _)
-      log.warn("Attempted to modify read-only look-up map")
-    end
-
-    household.group_to_coordinator_map.__newindex = function(_, _, _)
-      log.warn("Attempted to modify read-only look-up map")
-    end
-
-    private.households[id] = household
-
-    -- emit group info update [groupId, groupRole, groupPrimaryDeviceId]
-    if device ~= nil then
-      log.debug(string.format("Emitting group info update for Device %s", device.label))
-      local device_player_id = device:get_field(PlayerFields.PLAYER_ID)
-      local group_id = self:get_group_for_player(id, device_player_id)
-      local coordinator_id = self:get_coordinator_for_player(id, device_player_id)
-
-      local role
-      if device_player_id == coordinator_id then
-        if #household.groups[group_id].playerIds > 1 then
-          role = "primary"
-        else
-          role = "ungrouped"
-        end
-      else
-        role = "auxilary"
-      end
-
-      local group_update_payload = { role, coordinator_id, group_id }
-      if type(coordinator_id) == "string" and type(group_id) == "string" then
-        handlers.handle_group_update(device, group_update_payload)
-      else
-        log.warn(
-        st_utils.stringify_table(
-            { household = household, group_update_payload = group_update_payload },
-            "Household update with invalid data",
-            false
-          )
-        )
-      end
-    end
-  end
-
-  --- @param self SonosState
-  --- @param household_id HouseholdId
-  --- @param player_id PlayerId
-  --- @return GroupId?,string?
-  ret.get_group_for_player = function(self, household_id, player_id)
-    log.debug_with({ hub_logs = true },
-      st_utils.stringify_table(
-        { household_id = household_id, player_id = player_id }, "Get Group For Player", true
-      )
-    )
-    local household = private.households[household_id]
-    if household == nil then
-      log.error(st_utils.stringify_table({ household = household }, "Get group for invalid household", false))
-      return nil, st_utils.stringify_table({ household = household }, "Get group for invalid household", false)
-    end
-    return household.player_to_group_map[player_id]
-  end
-
-  --- @param self SonosState
-  --- @param household_id HouseholdId
-  --- @param group_id GroupId
-  --- @return PlayerId?,string?
-  ret.get_coordinator_for_group = function(self, household_id, group_id)
-    log.debug_with({ hub_logs = true },
-      st_utils.stringify_table(
-        { household_id = household_id, group_id = group_id }, "Get Coordinator For Group", true
-      )
-    )
-    local household = private.households[household_id]
-    if household == nil then
-      log.error(st_utils.stringify_table({ household = household }, "Get coordinator for invalid household", false))
-      return nil, st_utils.stringify_table({ household = household }, "Get coordinator for invalid household", false)
-    end
-    return household.group_to_coordinator_map[group_id]
-  end
-
-  --- @param self SonosState
-  --- @param household_id HouseholdId
-  --- @param player_id PlayerId
-  --- @return PlayerId?,string?
-  ret.get_coordinator_for_player = function(self, household_id, player_id)
-    log.debug_with({ hub_logs = true },
-      st_utils.stringify_table(
-        { household_id = household_id, player_id = player_id }, "Get Coordinator For Player", true
-      )
-    )
-    return self:get_coordinator_for_group(
-      household_id,
-      self:get_group_for_player(household_id, player_id)
-    )
-  end
-
-  --- @param self SonosState
-  --- @param device SonosDevice
-  --- @return HouseholdId|nil household_id nil on error
-  --- @return PlayerId|nil player_id nil on error
-  --- @return nil|string error nil on success
-  ret.get_player_for_device = function(self, device)
-    local household_id, player_id = device:get_field(PlayerFields.HOUSEHOULD_ID),
-      device:get_field(PlayerFields.PLAYER_ID)
-    log.debug_with({ hub_logs = true },
-      st_utils.stringify_table(
-        { name = (device or {label = "<no device>"}).label, household_id = household_id, player_id = player_id },
-        "Get Player For Device",
-        true
-      )
-    )
-
-    if not (device:get_field(PlayerFields._IS_INIT) or household_id or player_id) then
-      return nil, nil,
-          "Player not fully initialized: " .. device.label
-    end
-
-    return household_id, player_id, nil
-  end
-
-  --- @param self SonosState
-  --- @param device SonosDevice
-  --- @return HouseholdId|nil household_id nil on error
-  --- @return PlayerId|nil coordinator_id nil on error
-  --- @return nil|string error nil on success
-  ret.get_coordinator_for_device = function(self, device)
-    local household_id, player_id, err = self:get_player_for_device(device)
-    log.debug_with({ hub_logs = true },
-      st_utils.stringify_table(
-        { name = (device or {label = "<no device>"}).label, household_id = household_id, player_id = player_id },
-        "Get Coordinator For Device",
-        true
-      )
-    )
-    if err then
-      return nil, nil, err
-    end
-
-    local coordinator_id = self:get_coordinator_for_player(household_id, player_id)
-
-    if not coordinator_id then
-      return nil, nil, "Couldn't find coordinator ID for device " .. device.label
-    end
-
-    return household_id, coordinator_id, nil
-  end
-
-  --- @param self SonosState
-  --- @param device SonosDevice
-  --- @return HouseholdId|nil household_id nil on error
-  --- @return GroupId|nil group_id nil on error
-  --- @return nil|string error nil on success
-  ret.get_group_for_device = function(self, device)
-    local household_id, player_id, err = self:get_player_for_device(device)
-    log.debug_with({ hub_logs = true },
-      st_utils.stringify_table(
-        { name = (device or {label = "<no device>"}).label, household_id = household_id, player_id = player_id },
-        "Get Group For Device",
-        true
-      )
-    )
-    if err then
-      return nil, nil, err
-    end
-
-    local group_id = self:get_group_for_player(household_id, player_id)
-
-    if not group_id then
-      return nil, nil, "Couldn't find group ID for device " .. device.label
-    end
-
-    return household_id, group_id, nil
-  end
-
-  return ret
-end
-
-Types.SonosState = SonosState
-
---- @alias DiscoCallback fun(dni: string, ssdp_group_info: SonosSSDPInfo, player_info: SonosDiscoveryInfo, group_info: SonosGroupsResponseBody)
 
 ---@class SonosFieldCacheTable
 ---@field public swGen number
@@ -386,18 +179,10 @@ Types.SonosState = SonosState
 ---@field public player_id string
 ---@field public wss_url string
 
-
---- Sonos Edge Driver extensions
---- @class SonosDriver : Driver
---- @field public sonos SonosState Local state related to the sonos systems
---- @field public _player_id_to_device table<string,SonosDevice>
---- @field public _field_cache table<string,SonosFieldCacheTable>
---- @field public update_group_state fun(self: SonosDriver, header: SonosResponseHeader, body: SonosGroupsResponseBody)
---- @field public handle_ssdp_discovery fun(self: SonosDriver, ssdp_group_info: SonosSSDPInfo, callback?: DiscoCallback)
---- @field public is_same_mac_address fun(dni: string, other: string): boolean
-
 --- Sonos Player device
 --- @class SonosDevice : st.Device
+--- @field public id string
+--- @field public log table device-scoped logging module
 --- @field public label string Device label set at `try_create_device`
 --- @field public device_network_id string the DNI of the device
 --- @field public type string Network type. Should be "LAN".
@@ -407,8 +192,8 @@ Types.SonosState = SonosState
 --- @field public vendor_provided_label string The vendor's model string
 --- @field public get_field fun(self: SonosDevice, key: string):any
 --- @field public set_field fun(self: SonosDevice, key: string, value: any, args?: table)
+--- @field public emit_event fun(self: SonosDevice, event: any)
+--- @field public driver SonosDriver
 
 --- Sonos JSON commands
 --- @class SonosCommand
-
-return Types
